% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/changepoint.mv.class.R
\docType{methods}
\name{cpts.uv}
\alias{cpts.uv}
\alias{cpts.uv,changepoint.mv.mrc.class-method}
\title{Univariate changepoint locations.}
\usage{
\S4method{cpts.uv}{changepoint.mv.mrc.class}(x)
}
\arguments{
\item{x}{An S4 object as returned by \code{\link{mrc}}.}
}
\value{
A list of \eqn{N} vectors containing the univariate changepoint locations. Each vector corresponds to an individual variate in the data.
}
\description{
Returns a list of vectors containing the univariate changepoint locations.
}
\examples{
library(changepoint.mv)
data(mrcexample)
res<-mrc(mrcexample)
cpts.uv(res)

}
