% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/changepoint.mv.class.R
\docType{methods}
\name{show}
\alias{show}
\alias{show,changepoint.mv.mrc.class-method}
\title{Displays S4 objects produced by changepoint.mv methods}
\usage{
\S4method{show}{changepoint.mv.mrc.class}(object)
}
\arguments{
\item{object}{An S4 object produced by \code{\link{mrc}}.}
}
\description{
Displays S4 object produced by \code{\link{mrc}}. The information produced is the same as that provided by the \code{summary} method.
The method is used by the S4 system for automatic printing.
}
\examples{

library(changepoint.mv)
data(mrcexample)
res<-mrc(mrcexample)
# the following lines all produce the same output
res
summary(res)
show(res)
print(res)

}
