% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/changepoint.mv.class.R
\docType{methods}
\name{summary}
\alias{summary}
\alias{summary,changepoint.mv.class-method}
\alias{summary,changepoint.mv.mrc.class-method}
\alias{library(changepoint.mv)}
\alias{data(mrcexample)}
\alias{res<-spot(mrcexample)}
\alias{summary(res)}
\title{Summary}
\usage{
\S4method{summary}{changepoint.mv.class}(object)

\S4method{summary}{changepoint.mv.mrc.class}(object)
}
\arguments{
\item{object}{An S4 class instance produced by \code{\link{mrc}}.}
}
\description{
Produces and prints a summary of the results contained in an S4 class instance produced by \code{\link{mrc}}. The information
produced depends on the analysis type and options used, but typically includes the total penalised cost, penalty values and cost function type.
}
